// src/store/interfaces.ts
export interface IPreference {
  name: string;
  type: "text" | "boolean" | "range" | "list" | "hidden";
  label: string;
  options?: IPreferenceOption[];
  defaultValue?: string;
  value?: string;
}

/**
 * Interface for constructing user preference option
 *
 * @property label - preference displayed name
 * @property value - preference hidden value
 *
 */

export interface IPreferenceOption {
  label: string;
  value: string;
}

export interface IWidgetUserSettings {
  [id: string]: any;
}